requirements.txt
Flask
python-telegram-bot==13.15
pydub
ffmpeg-python
paho-mqtt
psycopg2-binary